django/requirements.txt
Django
gunicorn
mysqlclient
djangorestframework
langchain
openai
chromadb
tiktoken
unstructured
requests